ipvsadm: fix list_daemon to handle master/backup status in either position
Attached is a patch that fixes the list_daemon function such that it
does not assume that the master sync daemon status is always in the
first position and master sync daemon status is always in the second
position.

If libipvs uses the netlink interface to retrieve sync daemon status,
the results are not guaranteed to follow this ordering. As explained
in a previous email, if libipvs uses the netlink interface to retrieve
sync daemon status while only a backup sync daemon is running, the
backup sync daemon status will but in the first position (index 0).
This differs from the getsockopt interface, which would always put
master sync daemon status in first position and backup sync daemon
status in the second position, even when only backup sync daemon
exists. Solution is to make ipvsadm check both elements of the array
for master and backup.

Signed-off-by: Simon Horman <horms@verge.net.au>