ipvsadm: detect LIBS and CFLAGS for libnl versions
Automated detecting of LIBS and CFLAGS for different versions
of libnl via using pkg-config, in the Makefiles.  This should
make it easier to compile on different distributions.

Detecting in prioritized order from the most recent version, and only
the first detected lib is used. Thus, supporting several libnl version
being installed on the same system.

Only tested with libnl-1 and libnl-3.0, but also adding support for
libnl-2.0.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>
Acked-by: Daniel Borkmann <dborkman@redhat.com>
Signed-off-by: Simon Horman <horms@verge.net.au>