libipvs: CHECK_IPV4 and CHECK_PE macros should not call return
The CHECK_IPV4 and CHECK_PE macros will call return if a conditional
is true. This is a problem for any caller that has allocated memory
and needs to do cleanup before returning. This patch replaces the
return in both macros with a goto statement. As a result, each caller
needs a out_err label.

Signed-off-by: Ryan O'Hara <rohara@redhat.com>
Acked-by: Julian Anastasov <ja@ssi.bg>
Signed-off-by: Simon Horman <horms@verge.net.au>