Maintainer script for releasing tarballs
Adding a script (release-tarball.sh) for making a release
tarball, where we avoid including files not related to the git
repository, by using git archive to a temp dir (mktemp).

You need to create a git tag corrosponding with the version in the
file VERSION, before you can make a release.  For testing purposes,
its possible to give a commit id on the command line via "-c" option.

The script also creates a GPG signed file output of the uncompressed
tarball (which is needed by kernel.org upload utility).  A specific
GPGKEY ID can be specified on the command line via "-g" option.

The release tarballs and GPG signing files are placed in the
directory "release/".

Acked-by: Simon Horman <horms@verge.net.au>
Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>